delete media once tweet has been made

+++ subalert/subq.py
@@ -1,6 +1,7 @@
 from subalert.subtweet import Tweet
 import os
 
+
 class Queue:
     def __init__(self):
         self.items = []
@@ -31,7 +32,10 @@ class Queue:
                 counter += 1
                 print(f"\n##[ batch_all ({counter}) ]###\n{tweet}\n{media}\n------\n")
                 Tweet("NonFungibleTxs").alert(message=tweet, filename=media, verbose=True)
-                os.remove(path=media)
+
+                # only remove media if it actually returns anything.
+                if media:
+                    os.remove(path=media)
 
         if len(self.items[0]['transactions']) >= 1 and self.items[0]['transactions'][0] is not None:
             for tx in self.items[0]['transactions']:
